+++ src/main/java/com/app/HotelHungerGames/entity/BidEntity.java
@@ -3,6 +3,7 @@ package com.app.HotelHungerGames.entity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDateTime;
 
@@ -16,7 +17,7 @@ public class BidEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private AuctionEntity auction;
     private String email;
-    private Double price;
+    private BigDecimal price;
     private Instant bidTime;
 
 
@@ -24,7 +25,7 @@ public class BidEntity {
     }
 
     public BidEntity(Long id, AuctionEntity auction, String email,
-                     Double price, Instant bidTime) {
+                     BigDecimal price, Instant bidTime) {
         this.id = id;
         this.auction = auction;
         this.email = email;
@@ -56,11 +57,11 @@ public class BidEntity {
         this.email = email;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
